+++ package/freedombase-flashnews/package.js
@@ -1,7 +1,7 @@
 /* global Package */
 Package.describe({
   name: 'freedombase:flashnews',
-  version: '0.6.0',
+  version: '0.7.0',
   // Brief, one-line summary of the package.
   summary: 'Timed and localized flash messages for your Meteor app ',
   // URL to the Git repository containing the source code for this package.
@@ -12,10 +12,10 @@ Package.describe({
 })
 
 Package.onUse(function (api) {
-  api.versionsFrom(['2.8.1', '3.0-beta.0'])
-  api.use(['ecmascript', 'typescript', 'mongo', 'callback-hook'])
+  api.versionsFrom(['2.8.1', '3.0-rc.0'])
+  api.use(['ecmascript', 'typescript', 'mongo', 'ddp', 'callback-hook'])
   api.use([
-    'aldeed:collection2@4.0.0-beta.6',
+    'aldeed:collection2@3.4.1',
     'socialize:base-model@1.1.7'
   ])
   api.mainModule('common.ts', 'client')